Remove postinstall script (#54)
Having a postinstall script that runs typescript has proven to be problematic for multiple reasons. The first one being that not all projects use typescript, which forced us to use npx in our build script. The next problem to surface is that @types/node doesn't exist in the install since it's a dev dependency of this repo. I don't consider adding @types/node to the dependencies of this package to be a good solution, so I'd prefer to just unwind the problem by not having a postinstall script at all.

This just means that developers in this repo should now run the build script after cloning and installing it.