Update modifiedRequest.body before it's used by hooks
The v3 implementation of RESTDataSource did not have the concept of a "modified"
request. The params and headers were updated in place and the body was passed
directly through to the willSendRequest and resolveURL hooks. The body was then
updated if necessary, meaning neither of those hooks actually saw the final
outgoing request.

The updated implementation for v4 was indeed a regression, where the original
body was no longer included on the request object before those hooks were called.

To remediate this, let's fully build our modified request object (including final
updates to the body and headers) before we call any hooks. This resolves the
undefined body regression, and does a better job of reporting the actual outgoing
request to the willSendRequest and resolveURL hooks`.

This does not restore original v3 behavior, but it is not a breaking change
against v4 since it is strictly additive. There is a slight change in behavior in
that when the body is JSON stringified (from an array or object), the response
headers' content-type will be updated to reflect that if not already set. I would
consider this also additive.

Fixes #83