Strip undefined values from URL params; disallow some forms for `params` (#88)
Being able to specify `params` as a single record object is very
convenient, but it becomes inconvenient when some parameters are
optional. Previously you could not use the record object syntax if a
parameter is optional. Now you can use that syntax and we will strip
undefined values, like `JSON.stringify`.

Previously we allowed all of the various forms of data that can be
passed to `new URLSearchParams()` (and in fact just passed what we got
directly to there). Implementing this undefined-stripping for all of the
forms (like "arbitrary iterable object of two-element arrays") would be
a lot of code copied from the DOM implementation, but *not* stripping
undefined from the other forms would be inconsistent. So we change to
*only* allow the record form (which is the only form that our docs show)
plus providing `URLSearchParams` yourself; if you used one of the other
forms you can just add a `new URLSearchParams()` to your code.

See the changeset entry for more details.

Fixes #24.