Undo changes to `willSendRequest`, etc.
This commit undoes destructive changes to the semantics of `willSendRequest`
i.e. when it is called and with what parameters.

The original `body` is now passed through to the `augmentedRequest` as
soon as it is created and before calling `willSendRequest`, so the
body can now always be observed when the hook is called.

Additionally, `Buffer`s are now passed through (and tested against)