Preserve `body` on request object passed to `willSendRequest`; fix `string` and `Buffer` body handling (#89)
The v4 update introduced multiple regressions w.r.t. the `modifiedRequest` object that was added.

* `body` was no longer being added to the intermediary request object before calling `willSendRequest`
* `modifiedRequest.body` was never being set in the case that the incoming body was a `string` or `Buffer`

This change resolves both by reverting to what we previously had in v3 (preserving the properties
on the incoming request object). The types had to be updated accordingly. Due to how `path` is
now handled in v4, the `willSendRequest` signature has been updated to  `(path, request)`.