Detect and serialize objects with `null` prototype; provide `shouldJSONSerializeBody` hook to override behavior (#90)
Simplify the JSON serialization conditional to the inverse:
Serialize if...
* it's an array or plain object
* if it has a toJSON() property on it (and isn't a Buffer or FormData-like)

Test explicitly that FormData is passed through rather than
all non-plain-objects. Passing in a random class is unexpected
unless it's meant to be serialized with a toJSON function.

Provide `shouldJSONSerializeBody` as an overridable method: 

The logic to decide on JSON serialization was previously baked in
to RESTDataSource. In case customization is desired, we've provided
this overridable method as an escape hatch for other use cases.