Extend cache TTL for responses with last-modified
Some cache entries can be used even after they "expire", if they have a
response header like "etag" or "last-modified" that lets you run a
special HTTP request that can get a 304 response (to mean "reuse the
data you have" rather than needing to fully send it again). So we have
special logic to keep those responses in our cache for longer... but
that logic only looks for etag, not last-modified. So now we look for
both.

As part of this, add a bunch of related comments and documentation.

In order to make the test of this functionality fail without the fix, it
needs to use a KeyValueCache that actually observes TTL, so switch to
the one we use in apollo-server. Note that this reveals that some
existing tests were actually reading cache values after TTL which
resulted in some minor changes like `age: 0` instead of missing `age`.

Fixes #34.