Allow specifying http-cache-semantics options in a Request
Fixes #53. This is based on
https://github.com/apollographql/apollo-server/pull/4306 by @harleynuss.

While it's generally not great to base our API on another package's API,
`http-cache-semantics` has been pretty stable, and if it makes major
changes we can always major-bump this package.

Because a type from `http-cache-semantics` is now part of our API, this
makes its DefinitelyTyped package a non-dev dependency. While we're
there, I noticed that `@apollo/utils.keyvaluecache` had a strict
dependency which I changed to a caret.

I originally wanted to make `httpCacheSemanticsCachePolicyOptions` part
of `CacheOptions` (next to `ttl`) but that function gets called *after*
the request is made (with the response as an argument), which is too
late.