New throwIfResponseIsError hook
We previously removed `didReceiveResponse` which let you override the
main "is it an error or not" logic. This restores it in an
error-specific way. This means you can throw errors even on 200s if you
want (or vice versa).

This also changes the `errorFromResponse` hook to take options instead
of a single argument, in case you want the error to reflect more data.

One negative change is you can no longer prevent `parseBody` from being
called, though you can always override `parseBody` to do less.

Fixes #32.