Introduce public `fetch` method for reading complete responses; introduce new `head()` method (#110)
-- Add public `fetch` method --

Users previously had no well-defined way to access the complete response (i.e.
for header inspection). The public API of HTTP helper methods only returned the
parsed response body. A `didReceiveResponse` hook existed as an attempt to
solve this, but its semantics weren't well-defined, nor was it a type safe approach
to solving the problem.

The new `fetch` method allows users to "bypass" the convenience of the HTTP
helpers in order to construct their own full request and inspect the complete
response themselves.

-- Update default `cacheKeyFor` to include method --

In its previous form, `cacheKeyFor` only used the URL to calculate the cache
key. As a result, when `cacheOptions.ttl` was specified, the method was
ignored. This could lead to surprising behavior where a POST request's
response was cached and returned for a GET request (for example).

The default `cacheKeyFor` now includes the request method, meaning there
will now be distinct cache entries for a given URL per method.

-- Provide head() HTTP helper method --

Some REST APIs make use of HEAD requests. It seems reasonable for us to
provide this method as we do the others.

It's worth noting that the API differs from the other helpers. While bodies are expected/allowed for other requests, that is explicitly not the case for HEAD
requests. This method returns the request object itself rather than a parsed
body so that useful information can be extracted from the headers.